Replaced src/lib/Modal/Index.svelte by custom tailored modals for classic lightbox and gallery
